add attendance feature3

+++ SmartCampus.App/Services/Implementations/AttendanceService.cs
@@ -20,12 +20,53 @@ namespace SmartCampus.App.Services.Implementations
             
             if (!_validStatuses.Contains(dto.Status)) throw new ArgumentException($"Invalid status. Must be one of: {string.Join(", ", _validStatuses)}");           
             if (!await _attendanceRepo.StudentExists(dto.StudentId)) throw new InvalidOperationException($"Student with ID {dto.StudentId} not found.");
+           
             if (!await _attendanceRepo.CourseExists(dto.CourseId)) throw new InvalidOperationException($"Course with ID {dto.CourseId} not found.");
+            
+            // بتأكد انى الطالب مسجل اصلا  في الكورس
+            if (!await _attendanceRepo.IsStudentEnrolledInCourse(dto.StudentId, dto.CourseId))
+                throw new InvalidOperationException("Student is not enrolled in this course.");
+            // منع  تكرر لتسجيل الحضور لو الطالب ناصح يعنى
             if (await _attendanceRepo.AttendanceExists(dto.StudentId, dto.CourseId, dto.Date))
             {
                 throw new InvalidOperationException(
                     $"Attendance already marked for student {dto.StudentId} in course {dto.CourseId} on {dto.Date.Date:yyyy-MM-dd}");
             }
+            // Rule Prevent marking Absent if student is Excused on this date
+            if (dto.Status == "Absent")
+            {
+                bool excused = await _attendanceRepo.AttendanceExistsWithStatus(dto.StudentId, dto.CourseId, dto.Date, "Excused");
+                if (excused)
+                {
+                    throw new InvalidOperationException("Student is already excused on this date. Cannot mark as Absent.");
+                }
+            }
+
+            //  منع تسجيل حضور اكتر من 5 محاضرات في اليوم 
+            var dailyCount = await _attendanceRepo.GetDailyAttendanceCount(dto.StudentId, dto.Date);
+            if (dailyCount >= 5)
+            {
+                throw new InvalidOperationException("A student cannot attend more than 6 classes per day.");
+            }
+
+            // منع تسجيل الحضور بعد ما بيعدى  7 أيام 
+            if (dto.Date.Date < DateTime.Today.AddDays(-7))
+            {
+                throw new InvalidOperationException("You cannot mark attendance older than 7 days.");
+              
+            }
+            // منع تسجيل الحضور لتاريخ لسه مجااااش
+            if (dto.Date.Date > DateTime.UtcNow.Date)
+            {
+                throw new InvalidOperationException("Attendance cannot be marked for a future date.");
+            }
+
+            // منع اسجله متأخر ولسه السيشن مبدأتش اصلا منطقى يعنى 
+            if (dto.Status == "Late" && dto.Date > DateTime.Now)
+            {
+                throw new InvalidOperationException("Cannot mark a student as Late before the session time.");
+            }
+
             var attendance = new Attendance
             {
                 StudentId = dto.StudentId,
@@ -57,7 +98,26 @@ namespace SmartCampus.App.Services.Implementations
             DateTime? from,
             DateTime? to)
         {
+
+            // from > to   ممنوع تكون 
+            if (from.HasValue && to.HasValue && from > to)
+                throw new ArgumentException("From date cannot be greater than To date");
+
+            // التاكد انى فعلا الطالب ده موجود
+            if (studentId.HasValue && !await _attendanceRepo.StudentExists(studentId.Value))
+                throw new InvalidOperationException($"Student with ID {studentId.Value} not found");
+
+            // لازم اتأكد انى الكورس ده موجود
+            if (courseId.HasValue && !await _attendanceRepo.CourseExists(courseId.Value))
+                throw new InvalidOperationException($"Course with ID {courseId.Value} not found");
+
+            // مش هعمل فلتر لفتره اكبر من 6 شهور
+            if (from.HasValue && (DateTime.Now - from.Value).TotalDays > 180)
+                throw new InvalidOperationException("Cannot filter data older than 6 monthes");
+
+            //  السطر ده هو اللى هيعمل الفلتر فعلا هو اللى بعمل الكويرى فى  الدتا بيز 
             var attendances = await _attendanceRepo.FilterAttendances(studentId, courseId, from, to);
+          
 
             return attendances.Select(a => new AttendanceDto
             {
@@ -73,15 +133,46 @@ namespace SmartCampus.App.Services.Implementations
 
         public async Task<object> GetAttendanceSummaryAsync(int studentId, int? courseId = null)
         {
-            // Verify student exists
+            // التاكد من انى الطالب موجود
             if (!await _attendanceRepo.StudentExists(studentId))
             {
                 throw new InvalidOperationException($"Student with ID {studentId} not found.");
             }
 
+            // التاكد انى الكورس موجود
+            if (courseId.HasValue && !await _attendanceRepo.CourseExists(courseId.Value))
+            {
+                throw new InvalidOperationException($"Course with ID {courseId.Value} not found.");
+            }
+
+            // لازم الطالب يكون مسجل فى الكورس
+            if (courseId.HasValue)
+            {
+                if (!await _attendanceRepo.IsStudentEnrolledInCourse(studentId, courseId.Value))
+                    throw new InvalidOperationException("Student is not enrolled in this course.");
+            }
+
+           
             var attendances = await _attendanceRepo.FilterAttendances(studentId, courseId, null, null);
             var attendanceList = attendances.ToList();
 
+            //لو مفيش حضور اصلا
+            if (!attendanceList.Any())
+            {
+                return new
+                {
+                    StudentId = studentId,
+                    CourseId = courseId,
+                    TotalClasses = 0,
+                    PresentCount = 0,
+                    LateCount = 0,
+                    AbsentCount = 0,
+                    ExcusedCount = 0,
+                    AttendancePercentage = 0
+                };
+            }
+
+           
             var total = attendanceList.Count;
             var present = attendanceList.Count(a => a.Status == "Present");
             var late = attendanceList.Count(a => a.Status == "Late");
@@ -103,6 +194,7 @@ namespace SmartCampus.App.Services.Implementations
             };
         }
 
+
         public async Task UpdateAttendanceAsync(int id, string status)
         {
             // Validate status
@@ -114,7 +206,18 @@ namespace SmartCampus.App.Services.Implementations
             var existingAttendance = await _attendanceRepo.GetAttendanceById(id);
             if (existingAttendance == null)
             {
-                throw new InvalidOperationException($"Attendance record with ID {id} not found.");
+                throw new InvalidOperationException($"Attendance record with ID {id} not found");
+            }
+            //  No update after 48 hours
+            if ((DateTime.UtcNow - existingAttendance.Date).TotalHours > 48)
+            {
+                throw new InvalidOperationException("Cannot update attendance after 48 hours");
+            }
+
+            // Prevent updating with same status
+            if (existingAttendance.Status == status)
+            {
+                throw new InvalidOperationException("Status is already the same");
             }
 
             existingAttendance.Status = status;
@@ -123,11 +226,34 @@ namespace SmartCampus.App.Services.Implementations
 
         public async Task DeleteAttendanceAsync(int id)
         {
+            // Check attendance exists
+            var attendance = await _attendanceRepo.GetAttendanceById(id);
+
+
+            if (attendance == null)
+            {
+                throw new InvalidOperationException($"Attendance with ID {id} not found");
+            }
+
+            //Cannot delete records older than 7 days
+            if (attendance.Date.Date < DateTime.Today.AddDays(-7))
+            {
+                throw new InvalidOperationException("You cannot delete attendance older than 7 days");
+            }
+
+            //Cannot delete attendance from the future
+            if (attendance.Date.Date > DateTime.Today)
+            {
+                throw new InvalidOperationException("You cannot delete attendance for a future date");
+            }
+
+            
             var deleted = await _attendanceRepo.DeleteAttendance(id);
             if (!deleted)
             {
-                throw new InvalidOperationException($"Attendance record with ID {id} not found.");
+                throw new InvalidOperationException("Delete failed");
             }
         }
+
     }
 }
